+++ CHANGELOG.md
@@ -12,7 +12,10 @@
 
 - Relicensed from MIT/Apache-2.0 to BlueOak-1.0.0/MIT/Apache-2.0.
 
-- Increased the minimum supported version of Rust from 1.7.0 to 1.34.0.
+- Increased the minimum supported version of Rust from 1.7.0 to 1.36.0.
+
+- no_std is now possible in the usual way (default Cargo feature 'std'),
+  depending on alloc and hashbrown.
 
 - Removed the `bench` Cargo feature which was mostly to work around historical
   Cargo limitations, but was solved by moving benchmarks from `src/lib.rs` to